fix(server): Fixed use `accessToken` to check for vulnerabilities on the date of received
